+++ __init__.py
@@ -15,8 +15,13 @@ def unregister():
     mouse_strafing.unregister_keymaps()
     bpy.utils.unregister_class(prefs.MouseStrafingPreferences)
     bpy.utils.unregister_class(mouse_strafing.MouseStrafingOperator)
+    bpy.utils.unregister_class(mouse_strafing.CameraStates)
+    bpy.utils.unregister_class(mouse_strafing.CameraState)
 
 def register():
+    bpy.utils.register_class(mouse_strafing.CameraState)
+    bpy.utils.register_class(mouse_strafing.CameraStates)
+    bpy.types.Scene.mstrf_camera_save_states = bpy.props.PointerProperty(type = mouse_strafing.CameraStates, options = {"HIDDEN"})
     bpy.utils.register_class(mouse_strafing.MouseStrafingOperator)
     bpy.utils.register_class(prefs.MouseStrafingPreferences)
     mouse_strafing.register_keymaps()